requirements.txt
streamlit
pandas
numpy
matplotlib
tensorflow
google-generativeai
python-dotenv
scikit-learn